Изменены файлы начальной настройки бота
