+++ config.py
@@ -6,26 +6,7 @@ config_file = configparser.ConfigParser()
 config = config_file['DEFAULT']
 
 current_path = os.path.abspath(os.path.dirname(__file__))
-if os.path.exists(current_path + '/' + "config.ini"):
-    config_file.read(current_path + '/' + "config.ini")
-else:
-    config = {
-        'TOKEN': 'место для токена',
-        'WEBHOOK_HOST': '',
-        'WEBHOOK_PORT': '',
-        'WEBHOOK_LISTEN': '0.0.0.0',
-        'WEBHOOK_SSL_CERT': '',
-        'WEBHOOK_SSL_PRIV': '',
-        'DB_NAME': 'название базы данных',
-        'DB_HOST': 'адрес БД',
-        'DB_USER': 'пользователь для работы с БД',
-        'DB_PASSWORD': 'пароль пользователя',
-        'LOG_DIR_PATH': current_path + '/' + 'log' + '/',
-        'WEEK_TYPE': '0',
-        'STATISTIC_TOKEN': ''}
-    with open(current_path + '/' + 'config.ini', 'w') as configfile:
-        config_file.write(configfile)
-
+config_file.read(current_path + '/' + "config.ini")
 
 daysOfWeek = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"]
