+++ install/setup.py
@@ -1,21 +1,13 @@
 # -*- coding: utf-8 -*-
 import configparser
 import os
-import psycopg2
-
-
-def init_db(name, user, pasw, host, schema_path):
-    with psycopg2.connect(dbname=name, user=user, password=pasw, host=host) as db:
-        with open(schema_path, "r") as f:
-            db.cursor().execute(f.read())
-        db.commit()
 
 
 if __name__ == "__main__":
     # Считывание конфигурационного файла или его создание с настройками по умолчанию
     config = configparser.ConfigParser()
 
-    current_path = os.path.abspath(os.path.dirname(__file__))
+    current_path = os.path.abspath(os.path.dirname(__file__))[:-7]
     if os.path.exists(current_path + '/' + "config.ini"):
         config.read(current_path + '/' + "config.ini")
     else:
@@ -41,14 +33,3 @@ if __name__ == "__main__":
         os.makedirs(config["DEFAULT"]["LOG_DIR_PATH"])
     except OSError:
         pass
-
-    # Настройка базы данных
-    try:
-        init_db(
-            name=config["DEFAULT"]["DB_NAME"],
-            user=config["DEFAULT"]["DB_USER"],
-            pasw=config["DEFAULT"]["DB_PASSWORD"],
-            host=config["DEFAULT"]["DB_HOST"],
-            schema_path=current_path + "/" + "schema.sql")
-    except BaseException as e:
-        print(str(e))
